Create IAM_Roles
IAM roles is a simple script to run and create multiple roles as an organization deploying a project with various departments, the conditional expression uses the variable $EXPIRY_DATE to set the intended amount of access time the users can access the project of the organization. The script gives us an example of how to use conditional expressions while interacting with google cloud APIs for resources, users, and service accounts across products on the platform.